vector.ts renamed to vector2.ts
